modularizing maketile
